feat: change RN SDK API, remove context/provider, update example app
