+++ example/src/components/Identity.tsx
@@ -6,29 +6,32 @@ import {
   Button,
   ActivityIndicator,
 } from 'react-native';
-import { useStripeIdentity } from 'stripe-identity-react-native';
+import { Options, useStripeIdentity } from 'stripe-identity-react-native';
 
-export function Identity() {
-  const { status, present, loading } = useStripeIdentity();
+type Props = {
+  fetchOptions: () => Promise<Options>;
+};
+
+export function Identity({ fetchOptions }: Props) {
+  const { status, present, loading } = useStripeIdentity(fetchOptions);
 
   const handlePress = useCallback(() => {
     present();
   }, [present]);
 
-  const renderContent = useCallback(() => {
+  const renderButton = useCallback(() => {
     if (loading) {
       return <ActivityIndicator />;
     }
+    return <Button title="Verify Identity" onPress={handlePress} />;
+  }, [loading, handlePress]);
 
-    return (
-      <View style={styles.container}>
-        <Button title="Verify Identity" onPress={handlePress} />
-        <Text>Status: {status}</Text>
-      </View>
-    );
-  }, [loading, status, handlePress]);
-
-  return <View style={styles.container}>{renderContent()}</View>;
+  return (
+    <View style={styles.container}>
+      <View style={styles.buttonContainer}>{renderButton()}</View>
+      <Text>Status: {status}</Text>
+    </View>
+  );
 }
 
 const styles = StyleSheet.create({
@@ -36,5 +39,8 @@ const styles = StyleSheet.create({
     flex: 1,
     alignItems: 'center',
     justifyContent: 'center',
+  },
+  buttonContainer: {
+    height: 50,
   },
 });
