--- src/components/StripeIdentityContext.ts
@@ -1,14 +1,0 @@
-import { createContext } from 'react';
-import type { IdentityStatus } from '../types';
-
-type ContextType = {
-  status: IdentityStatus;
-  setStatus(status: IdentityStatus): void;
-  loading: boolean;
-};
-
-export const StripeIdentityContext = createContext<ContextType>({
-  status: 'Idle',
-  setStatus: () => {},
-  loading: true,
-});
